refactor: move supabase-package to singletons-package
